itas-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
PyPDF2==3.0.1
# spacy==3.7.2
nltk==3.8.1
psycopg2-binary
python-dotenv
Pillow>=10.3.0
